Fixed the average lines code
The call to polyfit on line 39 of the average_slope_intercept
mixed x1, y1, x2, y2. On looking at the comments on polyfit
it takes x and y so we should pass (x1, x2), (y1, y2) not
mix them as we did. Also make a change to make_coordinates to
choose y1 being a point on the line and y2 being 60% up the
screen. Changed video capture code to check for empty frame
at the end so that app does not crash on last frame